Move junit platform under junit package
