+++ subprojects/testing-jvm/src/integTest/groovy/org/gradle/testing/junit/platform/JUnitPlatformUserGuideIntegrationTest.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 the original author or authors.
+ * Copyright 2023 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.testing.junitplatform
+package org.gradle.testing.junit.platform
 
 import org.gradle.integtests.fixtures.DefaultTestExecutionResult
